Create a shell script version of 'pokemonsay'
I didn't like the script written in python. I feel like I didn't do
it quite right. The mess with pipes confuses me. It is so much
simpler to handle pipes using shell script, in my opnion.

The shell script version of 'pokemonsay' is tiny if compared to the
python version. And it does the same thing.