Create a install script
In order to use 'pokemonsay', I need it installed somewhere. I think
a simple script is enough for that.

The install script copies the essential files to the '.config' folder
in 'home', and creates a script in the '~/bin' directory that will
run 'pokemonsay'.

I wonder if it could have been done in any better way.