Embed the uninstall script inside the install script
A heredoc was used in the install script to generate the uninstall
script. This was made because they share important variables, like
'$install_path'.

If someone, for example, changes the '$install_path' in the install
script, but not in the uninstall script. The latter would be brokken.
So I think this was the safest way to keep things right, even if
changes are made.