Fix the bug ignore the user provided message
A bug was introduced in commit 1fc82b6, when word wrapping was
introduced. The script was ignoring the user provided message.
What made it hang waiting for user input.

The bug is fixed now. Who could imagine that just a pair of quote
marks could break the script like that...